Move GBuf and SSAO compute to separate nodes
