+++ renderer-gpu/src/lib.rs
@@ -30,6 +30,8 @@ use std::thread::spawn;
 use strum::IntoEnumIterator;
 use tokio::sync::broadcast;
 use wgpu::{Texture, TextureView};
+use crate::pass_nodes::g_buffer_pass_node::GBufferPassNode;
+use crate::pass_nodes::ssao_pass_node::SsaoPassNode;
 use crate::render_config::RenderConfig;
 
 pub mod canvas_api;
@@ -54,6 +56,8 @@ mod buffer_pool;
 
 pub mod wgpu_canvas;
 pub mod render_config;
+
+pub(crate) mod texture_view_resources;
 
 pub struct GpuRenderer {
     render_config: RenderConfig,
@@ -186,6 +190,7 @@ impl GpuRenderer {
                     .set_texture(texture_view, (-100.0, -100.0), &self.global_context, &mut self.buffer_pool);
             }
         }
+        self.config_pass_nodes();
     }
 
     fn config_pass_nodes(&mut self) {
@@ -193,6 +198,10 @@ impl GpuRenderer {
         let shadow_pass_node = ShadowPrepass::new();
 
         let rt_node = RenderToTexturePassNode::new(&mut self.global_context);
+
+        let g_buf_node = GBufferPassNode::new(&mut self.global_context);
+        let ssao_node = SsaoPassNode::new(&mut self.global_context);
+
         let main_node = MainPassNode::new(&mut self.global_context);
 
         PreviewType::iter().for_each(|preview_type| {
@@ -200,9 +209,9 @@ impl GpuRenderer {
                 PreviewType::None => None,
                 PreviewType::Camera => Some(rt_node.rt_texture_view.clone()),
                 PreviewType::SSAO => Some(self.global_context.ssao_texture.clone()),
-                PreviewType::SSAOPositions => Some(main_node.non_msaa_texture_view_positions.clone()),
-                PreviewType::SSAONormals => Some(main_node.non_msaa_texture_view_normals.clone()),
-                PreviewType::SSAODepth => Some(main_node.non_msaa_depth_texture_view.clone()),
+                PreviewType::SSAOPositions => self.global_context.texture_view_resources.non_msaa_texture_view_positions.clone(),
+                PreviewType::SSAONormals => self.global_context.texture_view_resources.non_msaa_texture_view_normals.clone(),
+                PreviewType::SSAODepth => self.global_context.texture_view_resources.non_msaa_depth_texture_view.clone(),
                 PreviewType::ShadowMap => Some(self.global_context.shadow_map_depth_texture.clone()),
             } {
                 self.preview_textures.insert(preview_type, texture_view);
@@ -220,9 +229,18 @@ impl GpuRenderer {
 
         self.pass_nodes = vec![Box::new(pre_pass_node)];
 
-        self.pass_nodes.push(Box::new(rt_node));
+        if self.render_config.preview_type != PreviewType::None {
+            self.pass_nodes.push(Box::new(rt_node));
+        }
 
-        self.pass_nodes.push(Box::new(shadow_pass_node));
+        if self.render_config.shadow_enabled {
+            self.pass_nodes.push(Box::new(shadow_pass_node));
+        }
+
+        if self.render_config.ssao_enabled {
+            self.pass_nodes.push(Box::new(g_buf_node));
+            self.pass_nodes.push(Box::new(ssao_node));
+        }
 
         self.pass_nodes.push(Box::new(main_node));
     }
@@ -264,11 +282,7 @@ impl GpuRenderer {
             });
 
         self.pass_nodes.iter_mut().for_each(|node| {
-            node.compute(&mut encoder,
-                         &mut self.layers,
-                         &mut self.global_context);
-
-            node.render(
+            node.run(
                 &mut encoder,
                 &mut self.layers,
                 &mut self.global_context,
