+++ renderer-gpu/src/mesh_layers/screen_shape_layer.rs
@@ -1,7 +1,7 @@
 use crate::buffer_pool::BufferPool;
 use crate::collider::{ColliderTask, CollisionTaskController, CollisionTaskWrapper};
 use crate::draw_commands::mesh2d_draw_command::Mesh2dDrawCommand;
-use crate::global_context::{GlobalContext, GlobalRenderStep};
+use crate::global_context::GlobalContext;
 use crate::mesh::InstanceBuffer;
 use crate::mesh::mesh::Mesh;
 use crate::mesh::mesh_instance_input::MeshInstanceInput;
