+++ renderer-gpu/src/pass_nodes/prepass_node.rs
@@ -1,8 +1,8 @@
 use crate::global_context::GlobalContext;
-use crate::mesh_layers::layers::Layers;
 use crate::mesh_layers::BaseMeshLayer;
+use crate::mesh_layers::layers::Layers;
 use crate::pass_nodes::PassNode;
-use wgpu::{CommandEncoder, TextureView};
+use wgpu::CommandEncoder;
 
 pub(crate) struct PrepassNode {}
 
@@ -13,21 +13,12 @@ impl PrepassNode {
 }
 
 impl PassNode for PrepassNode {
-    fn compute(
+    fn run(
         &self,
         encoder: &mut CommandEncoder,
         layers: &mut Layers,
         global_context: &mut GlobalContext,
     ) {
         layers.compute(encoder, global_context);
-    }
-
-    fn render(
-        &self,
-        _encoder: &mut CommandEncoder,
-        _layers: &mut Layers,
-        _global_context: &mut GlobalContext,
-    ) {
-        // no special rendering for prepass
     }
 }
